Add blocking in UserDataManager to make queue of requests

+++ packages/wordlist/src/controllers/user-data-manager.js
@@ -6,9 +6,13 @@ export default class UserDataManager {
 
   constructor (userID,events) {
     this.userID = userID
-    events.WORDITEM_UPDATED.sub(this.update.bind(this))
-    events.WORDITEM_DELETED.sub(this.delete.bind(this))
-    events.WORDLIST_DELETED.sub(this.deleteMany.bind(this))
+    if (events) {
+      events.WORDITEM_UPDATED.sub(this.update.bind(this))
+      events.WORDITEM_DELETED.sub(this.delete.bind(this))
+      events.WORDLIST_DELETED.sub(this.deleteMany.bind(this))
+    }
+    this.blocked = false
+    this.requestsQueue = []
   }
 
   _localStorageAdapter(dataType) {
@@ -50,15 +54,25 @@ export default class UserDataManager {
    * @return {Boolean} true if update succeeded false if not
    */
   async update(data) {
+    if (this.blocked) {
+      this.requestsQueue.push({
+        method: 'update',
+        data: data
+      })
+    }
     try {
+      this.blocked = true
       let finalConstrName = this.defineConstructorName(data.dataObj.constructor.name)
 
       let localAdapter = this._localStorageAdapter(finalConstrName)
       let remoteAdapter = this._remoteStorageAdapter(finalConstrName)
+
       let updatedLocal = await localAdapter.update(data.dataObj,data.params)
       let updatedRemote = await remoteAdapter.update(data.dataObj,data.params)
-      
       this.printErrors(localAdapter)
+      
+      this.blocked = false
+      this.checkRequestQueue()
 
       return updatedLocal && updatedRemote
     } catch (error) {
@@ -73,16 +87,26 @@ export default class UserDataManager {
    * @return {Boolean} true if delete succeeded false if not
    */
   async delete(data) {
+    if (this.blocked) {
+      this.requestsQueue.push({
+        method: 'delete',
+        data: data
+      })
+    }
     try {
+      this.blocked = true
       let finalConstrName = this.defineConstructorName(data.dataObj.constructor.name)
 
       let localAdapter = this._localStorageAdapter(finalConstrName)
       let remoteAdapter = this._remoteStorageAdapter(finalConstrName)
       let deletedLocal = await localAdapter.deleteOne(data.dataObj)
       let deletedRemote = await remoteAdapter.deleteOne(data.dataObj)
-
       this.printErrors(localAdapter)
-      // TODO error handling upon delete failure
+      
+      this.blocked = false
+
+      this.checkRequestQueue()
+      
       return deletedLocal && deletedRemote
     } catch (error) {
       console.error('Some errors happen on deleting data from IndexedDB', error.message)
@@ -97,16 +121,26 @@ export default class UserDataManager {
    *                      }
    */
   async deleteMany(data) {
+    if (this.blocked) {
+      this.requestsQueue.push({
+        method: 'deleteMany',
+        data: data
+      })
+    }
     try {
+      this.blocked = true
+
       let remoteAdapter =  this._remoteStorageAdapter(data.dataType)
       let localAdapter = this._localStorageAdapter(data.dataType)
       let deletedLocalResult = localAdapter.deleteMany(data.params)
       let deletedRemoteResult = remoteAdapter.deleteMany(data.params)
       const finalResult = [await deletedLocalResult, await deletedRemoteResult]
-
-      this.printErrors(localAdapter)
-
+      
+      this.printErrors(localAdapter)      
+      this.blocked = false
       console.info('Result of deleted many from IndexedDB', finalResult)
+
+      this.checkRequestQueue()
       
     } catch (error) {
       console.error('Some errors happen on deleting data from IndexedDB', error.message)
@@ -170,6 +204,13 @@ export default class UserDataManager {
       localAdapter.create(item)
     })
     return [...remoteDataItems,...addToRemote]
+  }
+
+  checkRequestQueue () {
+    if (this.requestsQueue.length > 0) {
+      let curRequest = this.requestsQueue.shift()
+      this[curRequest.method](curRequest.data)
+    }
   }
 
   printErrors (localAdapter) {
